+++ components_struct.html
@@ -1,0 +1,18 @@
+
+<Header></Header>
+<MainSection>
+    <Profile></Profile>
+    <RoomSection>
+        <h2></h2>
+        <ul className="Rooms">
+            <Room></Room>
+            <Room></Room>
+            <Room></Room>
+            //...
+        </ul>
+    </RoomSection>
+</MainSection>
+
+<ProfileModal>
+    
+</ProfileModal>
