+++ package.json
@@ -40,5 +40,5 @@
   "devDependencies": {
     "gh-pages": "^3.1.0"
   },
-  "homepage": "https://from97.github.io/pepple_front/"
+  "homepage": "https://from97.github.io/pepple/"
 }
