+++ src/components/AddNewRoomBtn.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+function AddNewRoomBtn() {
+    return (
+      <button className="AddNewRoom">방 새로 만들기</button>
+    );
+  }
+
+export default AddNewRoomBtn;
