+++ src/components/Header.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Header() {
+    return (
+      <header>
+        <h1>Pepple</h1>
+      </header>
+    );
+  }
+
+export default Header;
