+++ src/components/RoomSection.js
@@ -1,0 +1,26 @@
+import React from "react";
+import Room from "./Room";
+
+function RoomSection() {
+    return (
+      <div className="Room">
+        <h2>#Text</h2>
+        <ul className="room_list">
+          <Room />
+          <Room />
+          <Room />
+          <Room />
+          <Room />
+          <Room />
+          <Room />
+          <Room />
+          <Room />
+          <Room />
+          <Room />
+          <Room />
+        </ul>
+      </div>
+    )
+  }
+
+  export default RoomSection;
